+++ src/runners/episode_runner.py
@@ -221,8 +221,8 @@ class EpisodeRunner:
         return self.batch, self.macro_batch
 
     def _log(self, returns, stats, prefix):
-        self.logger.log_stat(prefix + "return_mean", np.mean(returns), self.t_env)
-        self.logger.log_stat(prefix + "return_std", np.std(returns), self.t_env)
+        self.logger.log_stat(prefix + "return_mean", float(np.mean(returns)), self.t_env)
+        self.logger.log_stat(prefix + "return_std", float(np.std(returns)), self.t_env)
         returns.clear()
 
         for k, v in stats.items():
